+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Custom difficulty recognised by the Galactic War Overhaul information panel
 - Nomad Commander's orbital structures can now move
 - Nomad Commander's structures now hover
+- Single Laser Defense Tower Upgrade will work correctly with the Nomad Commander loadout
 
 ## v4.13.1 - 2020-01-07